RiceChallenge_final/#3 requirements.txt
numpy==2.4.0
pandas==3.0.0
torch==2.10.0
torchvision==0.25.0
timm==1.0.24
albumentations==2.0.8
scikit-learn==1.8.0
rich==14.3.1
pillow==12.0.0
tqdm==4.67.1

